% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adhere_classr.R
\name{adhere_classr}
\alias{adhere_classr}
\title{Apply class for adherance data}
\usage{
adhere_classr(x)
}
\arguments{
\item{x}{data frame containing variables}
}
\description{
Automatically apply class to adherance data
}
\seealso{
\code{\link{tbcleanr}}
}
\author{
Jay Achar
}
