% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adhere_cleanr.default.R
\name{adhere_cleanr.default}
\alias{adhere_cleanr.default}
\title{Default method for adhere_cleanr()}
\usage{
\method{adhere_cleanr}{default}(x)
}
\arguments{
\item{x}{data frame containing variables}
}
\description{
Allow data frames with unspecified object class to pass through
}
\seealso{
\code{\link{tbcleanr}}
}
\author{
Jay Achar
}
