% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adhere_cleanr.epiinfo.R
\name{adhere_cleanr.epiinfo}
\alias{adhere_cleanr.epiinfo}
\title{Clean drug adherence data}
\usage{
\method{adhere_cleanr}{epiinfo}(x, adm = NULL)
}
\arguments{
\item{x}{data frame containing drug adherence data}

\item{adm}{optional arugement defining admission data frame to 
force function to nest adherence data.}
}
\description{
Take drug adherence data from routine TB programmes and clean for 
further analysis.
}
\seealso{
\code{\link{tbcleanr}}
}
\author{
Jay Achar
}
