% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adhere_cleanr.R
\name{adhere_cleanr}
\alias{adhere_cleanr}
\title{Clean drug adherence data}
\usage{
adhere_cleanr(x)
}
\arguments{
\item{x}{data frame containing drug adherence data}
}
\description{
Take drug adherence data from routine TB programmes and clean for 
further analysis.
}
\seealso{
\code{\link{tbcleanr}}
}
\author{
Jay Achar
}
