% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adherence_classr.R
\name{adherence_classr}
\alias{adherence_classr}
\title{Apply class for adherance data}
\usage{
adherence_classr(x)
}
\arguments{
\item{x}{data frame containing variables}
}
\description{
Automatically apply class to adherance data
}
\seealso{
\code{\link{tbcleanr}}
}
\author{
Jay Achar \email{jay.achar@doctors.org.uk}
}
