% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adm_binary_fixer.R
\name{adm_binary_fixer}
\alias{adm_binary_fixer}
\title{Convert binary variables to factorised output}
\usage{
adm_binary_fixer(x, set = "msc500", ...)
}
\arguments{
\item{x}{data frame containing drug variables}

\item{set}{define variable set to apply. Values can be "msc500",
"k6_adm_standard"}

\item{...}{further arguments passed to or from other methods}
}
\description{
Take data frame with binary variables and output 
binary factorised variables
}
\examples{
\dontrun{
adm_binary_fixer(p, set = "msc500")
}
}
\seealso{
\code{\link{TB.funs}}
}
\author{
Jay Achar \email{jay.achar@doctors.org.uk}
}
