% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adm_subset.R
\name{adm_subset.default}
\alias{adm_subset.default}
\title{Default method for adm_subset()}
\usage{
\method{adm_subset}{default}(x, add = NULL, ...)
}
\arguments{
\item{x}{data frame containing variables}

\item{add}{string of any additional variables to keep}

\item{...}{further arguments passed to or from other methods}
}
\description{
Allow data frames with unspecified object class to pass through
}
\seealso{
\code{\link{tbcleanr}}
}
\author{
Jay Achar
}
