% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adm_subset.epiinfo.R
\name{adm_subset.epiinfo}
\alias{adm_subset.epiinfo}
\title{Subset EpiInfo adm variables}
\usage{
\method{adm_subset}{epiinfo}(x, add = NULL, ...)
}
\arguments{
\item{x}{data frame containing variables}

\item{add}{string of any additional variables to keep}

\item{...}{further arguments passed to or from other methods}
}
\description{
Subset pre-specified EpiInfo TB admission variables
}
\seealso{
\code{\link{tbcleanr}}
}
\author{
Jay Achar \email{jay.achar@doctors.org.uk}
}
