% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adm_subset.koch6.R
\name{adm_subset.koch6}
\alias{adm_subset.koch6}
\title{Subset Koch 6 adm variables}
\usage{
\method{adm_subset}{koch6}(x, add = NULL, ...)
}
\arguments{
\item{x}{data frame containing variables}

\item{add}{string of any additional variables to keep}

\item{...}{further arguments passed to or from other methods}
}
\description{
Subset pre-specified Koch 6 TB admission variables
}
\seealso{
\code{\link{tbcleanr}}
}
\author{
Jay Achar \email{jay.achar@doctors.org.uk}
}
