% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/age_generator.R
\name{age_generator}
\alias{age_generator}
\title{Calculate age from two dates}
\usage{
age_generator(x, dob = "dateofbirth", start = "Starttre")
}
\arguments{
\item{x}{data frame containing Koch 6 admission variables}

\item{dob}{date of birth variable - "dateofbirth"}

\item{start}{date of treatment start - "Starttre"}
}
\description{
Use date of birth and date of starting treatment to calculate 
age at baseline. Takes data frame with date of birth and start 
of treatment date to generate age variable in years.
}
\examples{
\dontrun{
age_generator(p, dob = "dateofbirth", start = "Starttre")
}
}
\seealso{
\code{\link{TB.funs}}
}
\author{
Jay Achar \email{jay.achar@doctors.org.uk}
}
