% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apid_detangle.R
\name{apid_detangle}
\alias{apid_detangle}
\title{Detangle APID identification number into rayon, DS/DR TB and ID number}
\usage{
apid_detangle(x, rm_orig = TRUE, ...)
}
\arguments{
\item{x}{data frame containing APID variable from KK programme}

\item{rm_orig}{remove original variables - TRUE or FALSE}

\item{...}{further arguments passed to or from other methods}
}
\description{
Separate APID numbers into 3 constituent parts to ascertain 
district, DS or DR TB categorisation and numerical ID number
}
\examples{
\dontrun{
apid_detangle(p)
}
}
\seealso{
\code{\link{TB.funs}}
}
\author{
Jay Achar \email{jay.achar@doctors.org.uk}
}
