% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/baseline.R
\name{baseline}
\alias{baseline}
\title{Calculate baseline culture or smear status}
\usage{
baseline(x, id, result, date, start, neg, days = 90)
}
\arguments{
\item{x}{data frame or data.table with combined admission and lab data}

\item{id}{string variable representing patient unique ID number}

\item{result}{variable within data frame with laboratory result of interest. Must be a binary variable with missing values pre-removed.}

\item{date}{sample submission date - fomatted as date}

\item{start}{patient treatment start date - formatted as date}

\item{neg}{string representing factor level for negative sample result}

\item{days}{criteria for using pre-treatment samples}
}
\description{
This function takes a data frame and uses a unique patient identifier, dates of treatment start and sample submission and sample results to provide a baseline culture or smear result.
}
\keyword{TB}
