% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bentiu_xpert_clean.R
\name{bentiu_xpert_clean}
\alias{bentiu_xpert_clean}
\title{Bentiu Xpert laboratory data cleaning}
\usage{
bentiu_xpert_clean(x)
}
\arguments{
\item{x}{data frome including Bentiu PoC lab Xpert data}
}
\description{
This function takes the Bentiu PoC lab dataset as a data.frame and
removes unnecessary variables, cleans results and generates new 
result variables.
}
\examples{
\dontrun{
bentiu_xpert_clean(p)
}
}
\seealso{
\code{\link{TB.funs}}
}
\author{
Jay Achar \email{jay.achar@doctors.org.uk}
}
