% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binary_drug_fixer.R
\name{binary_drug_fixer}
\alias{binary_drug_fixer}
\title{Convert drug drug doses to binary output}
\usage{
binary_drug_fixer(x, set = "msc500")
}
\arguments{
\item{x}{data frame containing drug variables}

\item{set}{define variable set to apply. Values can be "msc500"}
}
\description{
Take data frame with drug dosing variables and output 
binary factorised variables
}
\examples{
\dontrun{
binary_drug_fixer(p, set = "msc500")
}
}
\seealso{
\code{\link{TB.funs}}
}
\author{
Jay Achar \email{jay.achar@doctors.org.uk}
}
