% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binary_fixer.R
\name{binary_fixer.default}
\alias{binary_fixer.default}
\title{Default method for binary_fixer()}
\usage{
\method{binary_fixer}{default}(x, ...)
}
\arguments{
\item{x}{data frame containing variables}

\item{...}{further arguments passed to or from other methods}
}
\description{
Allow data frames with unspecified object class to pass through
}
\seealso{
\code{\link{tbcleanr}}
}
\author{
Jay Achar
}
