% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binary_fixer.R
\name{binary_fixer}
\alias{binary_fixer}
\title{Convert admission binary variables to factorised output}
\usage{
binary_fixer(x, ...)
}
\arguments{
\item{x}{data frame containing admission binary variables}

\item{...}{further arguments passed to or from other methods}
}
\description{
Take admission data frame with binary variables and output 
binary factorised variables
}
\seealso{
\code{\link{tbcleanr}}
}
\author{
Jay Achar \email{jay.achar@doctors.org.uk}
}
