% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bmi_generator.R
\name{bmi_generator}
\alias{bmi_generator}
\title{Calculate BMI from height and weight}
\usage{
bmi_generator(x, weight = "weight", height = "height")
}
\arguments{
\item{x}{data frame containing height and weight variables}

\item{weight}{weight variable recorded in kgs - "weight"}

\item{height}{height variable recorded in cms - "height"}
}
\description{
Use numerical height and weight variables to calculate BMI
}
\examples{
\dontrun{
bmi_generator(p, weight = "weight", height = "height")
}
}
\seealso{
\code{\link{TB.funs}}
}
\author{
Jay Achar \email{jay.achar@doctors.org.uk}
}
