% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cavities_fixer.default.R
\name{cavities_fixer.default}
\alias{cavities_fixer.default}
\title{Convert admission x-ray variables to factorised output}
\usage{
\method{cavities_fixer}{default}(x, ...)
}
\arguments{
\item{x}{data frame containing admission x-ray variables}

\item{...}{further arguments passed to or from other methods}
}
\value{
admission data frame with factor x-ray summary varialbe
}
\description{
Convert admission x-ray variables to factorised output
}
\seealso{
\code{\link{tbcleanr}}
}
\author{
Jay Achar
}
