% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/change_classr.R
\name{change_classr}
\alias{change_classr}
\title{Add object class to identify change database}
\usage{
change_classr(x)
}
\arguments{
\item{x}{data frame containing raw TB change data}
}
\description{
From variable characteristics identify data entry tool and assign
object class to identify database for further cleaning
}
\seealso{
\code{\link{tbcleanr}}
}
\author{
Jay Achar \email{jay.achar@doctors.org.uk}
}
