% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/change_cleanr.R
\name{change_cleanr.default}
\alias{change_cleanr.default}
\title{Default method for change_cleanr()}
\usage{
\method{change_cleanr}{default}(x)
}
\arguments{
\item{x}{data frame containing drug variables}
}
\description{
Allow data frames with unspecified object class to pass through
}
\seealso{
\code{\link{tbcleanr}}
}
\author{
Jay Achar
}
