% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chechnya_lab_date_consolidator.R
\name{chechnya_lab_date_consolidator}
\alias{chechnya_lab_date_consolidator}
\title{Aggregate dates in Chechen lab database}
\usage{
chechnya_lab_date_consolidator(x, ...)
}
\arguments{
\item{x}{data frame containing sample date variables}

\item{...}{further arguments passed to or from other methods}
}
\description{
Consolidate sample dates in Chechen mycobacterial lab database
}
\examples{
\dontrun{
chechnya_lab_date_consolidator(p, set = "msc500")
}
}
\seealso{
\code{\link{TB.funs}}
}
\author{
Jay Achar \email{jay.achar@doctors.org.uk}
}
