% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_standard_indicators.R
\name{clean_standard_indicators}
\alias{clean_standard_indicators}
\title{Clean Annual Standard TB Indicators}
\usage{
clean_standard_indicators(file, year)
}
\arguments{
\item{file}{data frame of raw compiled project data.}

\item{year}{string defining year of data collection}
}
\value{
list containing a data frame of cleaned data and an integer indicating
the year of the raw data.
}
\description{
This function takes a data frame of imported data from  
a MSF OCA Annual Standard Indicators Excel file.
The function will clean, transpose, subset TB data and standardise renaming
missions and projects.
}
