% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/complete_tb_cleanr.R
\name{complete_tb_cleanr}
\alias{complete_tb_cleanr}
\title{Complete TB data clean}
\usage{
complete_tb_cleanr(adm, lab, software = c("excel", "koch_6", "epiinfo"),
  project = c("kk", "chechnya"), ...)
}
\arguments{
\item{adm}{raw admission data frame}

\item{lab}{raw laboratory data frame#'}

\item{software}{define software used for data collection.
Values can be "excel", "koch_6", "epiinfo"}

\item{project}{define project location to apply.
Values can be "kk", "chechnya".}

\item{...}{further arguments passed to or from other methods}
}
\description{
Combine all admission and laboratory data cleaning functions 
to organise TB admission and laboratory files into a nested list.
}
\seealso{
\code{\link{tbcleanr}}
}
\author{
Jay Achar \email{jay.achar@doctors.org.uk}
}
