% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/culture_recode.R
\name{culture_recode}
\alias{culture_recode}
\title{Recode culture results}
\usage{
culture_recode(cult)
}
\arguments{
\item{cult}{culture variable as defined in result_consolidator()}
}
\description{
Recode individual culture variables
}
\examples{
\dontrun{
culture_recode(cult)
}
}
\seealso{
\code{\link{TB.funs}}
}
\author{
Jay Achar \email{jay.achar@doctors.org.uk}
}
