% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/date.format.R
\name{date.format}
\alias{date.format}
\title{Detect and convert date variables}
\usage{
date.format(x, format = dmy)
}
\arguments{
\item{x}{data frame or data.table}

\item{format}{specify the lubridate output format}
}
\description{
This function detects variables with dates xx/xx/xxxx and converts them into date objects
}
\keyword{TB}
