% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/date_format.R
\name{date_format}
\alias{date_format}
\title{Detect and convert date variables}
\usage{
date_format(x, format = dmy, ...)
}
\arguments{
\item{x}{data frame or data.table}

\item{format}{specify the lubridate output format}

\item{...}{further arguments passed to or from other methods}
}
\description{
This function detects variables with dates xx/xx/xxxx and converts them into date objects
}
\keyword{TB}
