% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{drug_adhere}
\alias{drug_adhere}
\title{Drug adherence}
\usage{
drug_adhere(df, drug)
}
\arguments{
\item{df}{data frame with EpiInfo "Follow" data including all monthly adherence data}

\item{drug}{string representing drug to be analysed}
}
\description{
Calculate monthly drug adherence percentage
}
