% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drug_fixer.R
\name{drug_fixer.default}
\alias{drug_fixer.default}
\title{Default method for drug_fixer()}
\usage{
\method{drug_fixer}{default}(x, ...)
}
\arguments{
\item{x}{data frame containing variables}

\item{...}{further arguments passed to or from other methods}
}
\description{
Allow data frames with unspecified object class to pass through
}
\seealso{
\code{\link{tbcleanr}}
}
\author{
Jay Achar
}
