% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drug_fixer.epiinfo.R
\name{drug_fixer.epiinfo}
\alias{drug_fixer.epiinfo}
\title{Fix EpiInfo drug dosing data}
\usage{
\method{drug_fixer}{epiinfo}(x, ...)
}
\arguments{
\item{x}{data frame containing outcome variables}

\item{...}{further arguments passed to or from other methods}
}
\description{
Take data frame with drug dosing variables and output 
binary factorised variables
}
\seealso{
\code{\link{tbcleanr}}
}
\author{
Jay Achar \email{jay.achar@doctors.org.uk}
}
