% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drug_fixer.R
\name{drug_fixer}
\alias{drug_fixer}
\title{Convert drug doses to binary output}
\usage{
drug_fixer(x, ...)
}
\arguments{
\item{x}{data frame containing drug variables}

\item{...}{further arguments passed to or from other methods.}
}
\description{
Take data frame with drug dosing variables and output 
binary factorised variables
}
\seealso{
\code{\link{tbcleanr}}
}
\author{
Jay Achar \email{jay.achar@doctors.org.uk}
}
