% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dst_consolidator.R
\name{dst_consolidator.default}
\alias{dst_consolidator.default}
\title{Default method for dst_consolidator()}
\usage{
\method{dst_consolidator}{default}(x, ...)
}
\arguments{
\item{x}{data frame containing variables}

\item{...}{further arguments passed to or from other methods}
}
\description{
Allow data frames with unspecified object class to pass through
}
\seealso{
\code{\link{tbcleanr}}
}
\author{
Jay Achar \email{jay.achar@doctors.org.uk}
}
