% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dst_consolidator.R
\name{dst_consolidator}
\alias{dst_consolidator}
\title{Consolidate DST results}
\usage{
dst_consolidator(x, software = c("excel", "koch_6", "epiinfo"),
  project = c("kk", "chechnya"), file = c("adm", "lab", "clinical_lab"),
  aggregate = TRUE, rm_orig = TRUE, ...)
}
\arguments{
\item{x}{data frame containing variables}

\item{software}{define software used for data collection.
Values can be "excel", "koch_6", "epiinfo"}

\item{project}{define project location to apply.
Values can be "kk", "chechnya".}

\item{file}{define database file argument to apply.
Values can be "adm", "lab", "clinical_lab",}

\item{aggregate}{choose whether to aggregate to categories or retain all drug results}

\item{rm_orig}{remove original variables - TRUE or FALSE}

\item{...}{further arguments passed to or from other methods}
}
\description{
Take Chechen laboratory data set and consolidate DST results - choose aggregate
for categorised results
}
\examples{
\dontrun{
dst_consolidator(p, set = "chechnya_myco_lab", aggregate = TRUE, rm_orig = TRUE)
}
}
\seealso{
\code{\link{tbcleanr}}
}
\author{
Jay Achar \email{jay.achar@doctors.org.uk}
}
