% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dst_recoder.R
\name{dst_recoder}
\alias{dst_recoder}
\title{Recode DST results}
\usage{
dst_recoder(dst_var)
}
\arguments{
\item{dst_var}{DST variable}
}
\description{
Recode individual DST variables
}
\examples{
\dontrun{
dst_recoder(dst_var)
}
}
\seealso{
\code{\link{TB.funs}}
}
\author{
Jay Achar \email{jay.achar@doctors.org.uk}
}
