% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dstno_detangle.R
\name{dstno_detangle}
\alias{dstno_detangle}
\title{Detangle DST identification number}
\usage{
dstno_detangle(x, ...)
}
\arguments{
\item{x}{data frame containing APID variable from KK programme}

\item{...}{further arguments passed to or from other methods}
}
\description{
Convert character DST number found in Chechen data to numerical
}
\examples{
\dontrun{
dstno_detangle(p)
}
}
\seealso{
\code{\link{tbcleanr}}
}
\author{
Jay Achar \email{jay.achar@doctors.org.uk}
}
