% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epi_info_misc_cleanr.R
\name{epi_info_misc_cleanr}
\alias{epi_info_misc_cleanr}
\title{Claen miscellaneous Epi Info TB variables}
\usage{
epi_info_misc_cleanr(x, ...)
}
\arguments{
\item{x}{data frame containing Koch 6 admission variables}

\item{...}{further arguments passed to or from other methods}
}
\description{
Take data frame of TB admission data from Epi Info
and clean specific miscellaneous variables
}
\examples{
\dontrun{
epi_info_misc_cleanr(p)
}
}
\seealso{
\code{\link{tbcleanr}}
}
\author{
Jay Achar \email{jay.achar@doctors.org.uk}
}
