% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{extract_tb_data}
\alias{extract_tb_data}
\title{Extract TB data from imported Standard Indicators object}
\usage{
extract_tb_data(x)
}
\arguments{
\item{x}{data frame of imported Standard Indicators data}
}
\value{
data frame
}
\description{
Extract TB data from imported Standard Indicators object
}
