% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gender_fixer.default.R
\name{gender_fixer.default}
\alias{gender_fixer.default}
\title{Default method for gender_fixer()}
\usage{
\method{gender_fixer}{default}(x, ...)
}
\arguments{
\item{x}{data frame containing variables}

\item{...}{further arguments passed to or from other methods}
}
\description{
Allow data frames with unspecified object class to pass through
}
\seealso{
\code{\link{tbcleanr}}
}
\author{
Jay Achar \email{jay.achar@doctors.org.uk}
}
