% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gender_fixer.epiinfo.R
\name{gender_fixer.epiinfo}
\alias{gender_fixer.epiinfo}
\title{Factorise Epiinfo gender variable}
\usage{
\method{gender_fixer}{epiinfo}(x, ...)
}
\arguments{
\item{x}{data frame containing Koch 6 admission variables}

\item{...}{further arguments passed to or from other methods}
}
\description{
Take gender variable from Epiinfo data frame and factorise.
}
\seealso{
\code{\link{tbcleanr}}
}
\author{
Jay Achar \email{jay.achar@doctors.org.uk}
}
