% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gender_fixer.koch6.R
\name{gender_fixer.koch6}
\alias{gender_fixer.koch6}
\title{Factorise Koch 6 gender variable}
\usage{
\method{gender_fixer}{koch6}(x, ...)
}
\arguments{
\item{x}{data frame containing Koch 6 admission variables}

\item{...}{further arguments passed to or from other methods}
}
\description{
Take gender variable from Koch 6 data frame and factorise.
}
\seealso{
\code{\link{tbcleanr}}
}
\author{
Jay Achar \email{jay.achar@doctors.org.uk}
}
