% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hiv_fixer.epiinfo.R
\name{hiv_fixer.epiinfo}
\alias{hiv_fixer.epiinfo}
\title{Factorise Epiinfo HIV variable}
\usage{
\method{hiv_fixer}{epiinfo}(x, ...)
}
\arguments{
\item{x}{data frame containing EpiInfo admission variables}

\item{...}{further arguments passed to or from other methods}
}
\description{
Take HIV variables from data frame and factorise.
}
\seealso{
\code{\link{tbcleanr}}
}
\author{
Jay Achar \email{jay.achar@doctors.org.uk}
}
