% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hiv_fixer.koch6.R
\name{hiv_fixer.koch6}
\alias{hiv_fixer.koch6}
\title{Factorise Koch 6 HIV variable}
\usage{
\method{hiv_fixer}{koch6}(x, ...)
}
\arguments{
\item{x}{data frame containing Koch 6 admission variables}

\item{...}{further arguments passed to or from other methods}
}
\description{
Take HIV variables from data frame and factorise.
}
\seealso{
\code{\link{tbcleanr}}
}
\author{
Jay Achar \email{jay.achar@doctors.org.uk}
}
