% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/id_detangle.R
\name{id_detangle.default}
\alias{id_detangle.default}
\title{Default method for id_detangle()}
\usage{
\method{id_detangle}{default}(x, ...)
}
\arguments{
\item{x}{data frame containing admission ID variable}

\item{...}{further arguments passed to or from other methods}
}
\description{
Allow data frames with unspecified object class to pass through
}
\seealso{
\code{\link{tbcleanr}}
}
\author{
Jay Achar
}
