% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/id_detangle.epiinfo.R
\name{id_detangle.epiinfo}
\alias{id_detangle.epiinfo}
\title{Split admission ID number}
\usage{
\method{id_detangle}{epiinfo}(x, ...)
}
\arguments{
\item{x}{data frame containing APID variable from KK programme}

\item{...}{further arguments passed to or from other methods}
}
\description{
Handle TB data ID number based on database. If APID, split 
into 3 constituent parts to ascertain district, DS or DR TB 
categorisation and numerical ID number. If
}
\seealso{
\code{\link{tbcleanr}}
}
\author{
Jay Achar \email{jay.achar@doctors.org.uk}
}
