% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_standard_indicators.R
\name{import_standard_indicators}
\alias{import_standard_indicators}
\title{Custom import Annual Standard Indicators}
\usage{
import_standard_indicators(file)
}
\arguments{
\item{file}{character vector identifying the path and filename of an Excel 
file of MSF OCA Standard Indicators}
}
\value{
list containing a data frame of cleaned data and an integer indicating
the year of the raw data.
}
\description{
This function takes a single character vector identifying 
a MSF OCA Annual Standard Indicators Excel file of medical data.
The function will import, transpose, subset TB data and standardise renaming
missions and projects.
}
