% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/k6.clean.adm.R
\name{k6.clean.adm}
\alias{k6.clean.adm}
\title{Clean Koch6 admission data frame}
\usage{
k6.clean.adm(x, format = dmy)
}
\arguments{
\item{x}{data frame or data.table}

\item{format}{specify the lubridate output format}
}
\description{
This function takes a data frame derived from a Koch 6 admission export and formats multiple variables to manage dates and factors, while renaming.
}
\keyword{TB}
