% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/k6_adm_subset_vars.R
\name{k6_adm_subset_vars}
\alias{k6_adm_subset_vars}
\title{Subset Koch 6 admission variables}
\usage{
k6_adm_subset_vars(x, set = "msc500")
}
\arguments{
\item{x}{data frame containing Koch 6 admission variables}

\item{set}{define variable set to apply. Values can be "msc500"}
}
\description{
Subset Koch 6 admission variables along pre-defined variable  
sets
}
\examples{
\dontrun{
k6_adm_subset_vars(p, set = "msc500")
}
}
\seealso{
\code{\link{TB.funs}}
}
\author{
Jay Achar \email{jay.achar@doctors.org.uk}
}
