% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/k6_hep_fixer.R
\name{k6_hep_fixer}
\alias{k6_hep_fixer}
\title{Fix Koch 6 baseline viral hepatitis variables}
\usage{
k6_hep_fixer(x)
}
\arguments{
\item{x}{data frame containing Koch 6 HBV and HCV variables}
}
\description{
Combine Koch 6 HBV and HCV baseline variables and output  
integer varialbe
}
\examples{
\dontrun{
k6_hep_fixer(p)
}
}
\seealso{
\code{\link{TB.funs}}
}
\author{
Jay Achar \email{jay.achar@doctors.org.uk}
}
