% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/k6_idno_fixer.R
\name{k6_idno_fixer}
\alias{k6_idno_fixer}
\title{Fix Koch 6 registration number variable}
\usage{
k6_idno_fixer(x, ...)
}
\arguments{
\item{x}{data frame containing Koch 6 registration number}

\item{...}{further arguments passed to or from other methods}
}
\description{
Deconstruct Koch 6 registration number into constiuent parts 
and output numerical variable for idno and string for identifier, 
if relevant
}
\examples{
\dontrun{
k6_idno_fixer(p)
}
}
\seealso{
\code{\link{TB.funs}}
}
\author{
Jay Achar \email{jay.achar@doctors.org.uk}
}
