% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/k6_outcome_fixer.R
\name{k6_outcome_fixer}
\alias{k6_outcome_fixer}
\title{Fix Koch 6 treatment outcomes}
\usage{
k6_outcome_fixer(x)
}
\arguments{
\item{x}{data frame containing Koch 6 outcome variables}
}
\description{
Combine Koch 6 treatment outcome variables to leave one 
factorised, labelled variable
}
\examples{
\dontrun{
k6_outcome_fixer(p)
}
}
\seealso{
\code{\link{TB.funs}}
}
\author{
Jay Achar \email{jay.achar@doctors.org.uk}
}
