% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lab_classr.R
\name{lab_classr}
\alias{lab_classr}
\title{Add object class to identify lab database}
\usage{
lab_classr(x)
}
\arguments{
\item{x}{data frame containing raw TB admission data}
}
\description{
From variable characteristics identify data entry tool and assign
object class to identify database for further cleaning
}
\seealso{
\code{\link{tbcleanr}}
}
\author{
Jay Achar \email{jay.achar@doctors.org.uk}
}
