% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lab_data_cleanr.R
\name{lab_data_cleanr}
\alias{lab_data_cleanr}
\title{Cleans laboratory data sets}
\usage{
lab_data_cleanr(x, add = NULL, ...)
}
\arguments{
\item{x}{data frame containing variables}

\item{add}{string of any additional variables to keep}

\item{...}{further arguments passed to or from other methods}
}
\description{
Take laboratory data set and perform multiple adjustments based on which
laboratory data set is being used
}
\seealso{
\code{\link{tbcleanr}}
}
\author{
Jay Achar \email{jay.achar@doctors.org.uk}
}
