% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lab_date_consolidator.R
\name{lab_date_consolidator}
\alias{lab_date_consolidator}
\title{Aggregate dates in lab database}
\usage{
lab_date_consolidator(x, db, ...)
}
\arguments{
\item{x}{data frame containing sample date variables}

\item{db}{define database being used - "chechnya_myco_lab", "nukus_myco_lab"}

\item{...}{further arguments passed to or from other methods}
}
\description{
Consolidate sample dates in mycobacterial lab database
}
\examples{
\dontrun{
lab_date_consolidator(p, set = "msc500")
}
}
\seealso{
\code{\link{tbcleanr}}
}
\author{
Jay Achar \email{jay.achar@doctors.org.uk}
}
