% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lab_date_fixer.grozny.R
\name{lab_date_fixer.grozny}
\alias{lab_date_fixer.grozny}
\title{Aggregate dates in lab database}
\usage{
\method{lab_date_fixer}{grozny}(x, ...)
}
\arguments{
\item{x}{data frame containing sample date variables}

\item{...}{further arguments passed to or from other methods}
}
\description{
Consolidate sample dates in mycobacterial lab database
}
\seealso{
\code{\link{tbcleanr}}
}
\author{
Jay Achar \email{jay.achar@doctors.org.uk}
}
