% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lab_date_fixer.koch6.R
\name{lab_date_fixer.koch6}
\alias{lab_date_fixer.koch6}
\title{Aggregate dates in lab database}
\usage{
\method{lab_date_fixer}{koch6}(x, ...)
}
\arguments{
\item{x}{data frame containing sample date variables}

\item{...}{further arguments passed to or from other methods}
}
\description{
Consolidate sample dates in mycobacterial lab database
}
\seealso{
\code{\link{tbcleanr}}
}
\author{
Jay Achar
}
