% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lab_import_checkr.R
\name{lab_import_checkr.default}
\alias{lab_import_checkr.default}
\title{Default method for lab_import_checkr()}
\usage{
\method{lab_import_checkr}{default}(x, ...)
}
\arguments{
\item{x}{data frame containing variables}

\item{...}{further arguments passed to or from other methods}
}
\description{
Allow data frames with unspecified object class to pass through
}
\seealso{
\code{\link{tbcleanr}}
}
\author{
Jay Achar \email{jay.achar@doctors.org.uk}
}
