% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lab_import_checkr.epiinfo.R
\name{lab_import_checkr.epiinfo}
\alias{lab_import_checkr.epiinfo}
\title{Confirm correct lab data import}
\usage{
\method{lab_import_checkr}{epiinfo}(x)
}
\arguments{
\item{x}{data frame containing variables}
}
\value{
unchanged data frame or error highlighting parsing errors on data import
}
\description{
Confirm correct lab data import
}
\seealso{
\code{\link{tbcleanr}}
}
\author{
Jay Achar \email{jay.achar@doctors.org.uk}
}
