% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lab_import_checkr.koch6.R
\name{lab_import_checkr.koch6}
\alias{lab_import_checkr.koch6}
\title{Confirm correct lab data import}
\usage{
\method{lab_import_checkr}{koch6}(x)
}
\arguments{
\item{x}{data frame containing variables}
}
\value{
unchanged data frame or error highlighting parsing errors on data import
}
\description{
Confirm correct lab data import
}
\seealso{
\code{\link{tbcleanr}}
}
\author{
Jay Achar
}
