% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lab_import_checkr.R
\name{lab_import_checkr}
\alias{lab_import_checkr}
\title{Confirm correct lab data import}
\usage{
lab_import_checkr(x)
}
\arguments{
\item{x}{data frame containing variables}
}
\description{
By default, readr parses the first 1000 rows to define column class.
However, lab data is frequently very long with more recent test types, 
e.g. Hain MTBDRsl, only being available in more recent years. Thus, if not 
specified within the guess_max option, variables may be miss classified on import
}
\seealso{
\code{\link{tbcleanr}}
}
\author{
Jay Achar \email{jay.achar@doctors.org.uk}
}
