% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lab_longr.R
\name{lab_longr}
\alias{lab_longr}
\title{Convert lab data from wide to long}
\usage{
lab_longr(x, software = c("excel", "koch_6", "epiinfo"),
  project = c("kk", "chechnya"), file = c("adm", "lab",
  "clinical_lab"), add = NULL, ...)
}
\arguments{
\item{x}{data frame containing variables}

\item{software}{define software used for data collection.
Values can be "excel", "koch_6", "epiinfo"}

\item{project}{define project location to apply.
Values can be "kk", "chechnya".}

\item{file}{define database file argument to apply.
Values can be "adm", "lab", "clinical_lab",}

\item{add}{string of any additional variables to include}

\item{...}{further arguments passed to or from other methods}
}
\description{
In specified data sets, convert laboratory results from wide to long format.
}
\examples{
\dontrun{
lab_longr(z, set = "k6_clin_lab")
}
}
\seealso{
\code{\link{tbcleanr}}
}
\author{
Jay Achar \email{jay.achar@doctors.org.uk}
}
