% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lab_sample_fixer.grozny.R
\name{lab_sample_fixer.grozny}
\alias{lab_sample_fixer.grozny}
\title{Lab sample fixer}
\usage{
\method{lab_sample_fixer}{grozny}(x, rm_orig = TRUE, ...)
}
\arguments{
\item{x}{data frame containing sample date variables}

\item{rm_orig}{remove original variables - TRUE or FALSE}

\item{...}{further arguments passed to or from other methods}
}
\description{
Consolidate mycobacterial sample types in Chechen lab data
}
\seealso{
\code{\link{tbcleanr}}
}
\author{
Jay Achar \email{jay.achar@doctors.org.uk}
}
