% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lab_subset.grozny.R
\name{lab_subset.grozny}
\alias{lab_subset.grozny}
\title{Subset lab variables}
\usage{
\method{lab_subset}{grozny}(x, add = NULL, ...)
}
\arguments{
\item{x}{data frame containing variables}

\item{add}{string of any additional variables to keep}

\item{...}{further arguments passed to or from other methods}
}
\description{
Subset pre-specified TB laboratory variables and clean dstnumber variable
}
\seealso{
\code{\link{tbcleanr}}
}
\author{
Jay Achar \email{jay.achar@doctors.org.uk}
}
