% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lab_subset.koch6.R
\name{lab_subset.koch6}
\alias{lab_subset.koch6}
\title{Subset lab variables}
\usage{
\method{lab_subset}{koch6}(x, add = NULL, ...)
}
\arguments{
\item{x}{data frame containing variables}

\item{add}{string of any additional variables to keep}

\item{...}{further arguments passed to or from other methods}
}
\description{
Subset pre-specified TB laboratory variables
}
\seealso{
\code{\link{tbcleanr}}
}
\author{
Jay Achar
}
