% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lab_subset.R
\name{lab_subset}
\alias{lab_subset}
\title{Subset lab variables}
\usage{
lab_subset(x, add = NULL, ...)
}
\arguments{
\item{x}{data frame containing variables}

\item{add}{string of any additional variables to keep}

\item{...}{further arguments passed to or from other methods}
}
\description{
Subset pre-specified TB laboratory variables
}
\seealso{
\code{\link{tbcleanr}}
}
\author{
Jay Achar \email{jay.achar@doctors.org.uk}
}
