% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mtbdrplus_fixer.grozny.R
\name{mtbdrplus_fixer.grozny}
\alias{mtbdrplus_fixer.grozny}
\title{Consolidate Hain MTBDRplus results}
\usage{
\method{mtbdrplus_fixer}{grozny}(x, ...)
}
\arguments{
\item{x}{data frame containing variables}

\item{...}{further arguments passed to or from other methods}
}
\description{
Take laboratory data set and consolidate Hain MTBDRplus results
}
\seealso{
\code{\link{tbcleanr}}
}
\author{
Jay Achar \email{jay.achar@doctors.org.uk}
}
