% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mtbdrplus_fixer.R
\name{mtbdrplus_fixer}
\alias{mtbdrplus_fixer}
\title{Consolidate Hain MTBDRplus results}
\usage{
mtbdrplus_fixer(x, software = c("excel", "koch_6", "epiinfo"),
  project = c("kk", "chechnya"), file = c("adm", "lab", "clinical_lab"),
  ...)
}
\arguments{
\item{x}{data frame containing variables}

\item{software}{define software used for data collection.
Values can be "excel", "koch_6", "epiinfo"}

\item{project}{define project location to apply.
Values can be "kk", "chechnya".}

\item{file}{define database file argument to apply.
Values can be "adm", "lab", "clinical_lab",}

\item{...}{further arguments passed to or from other methods}
}
\description{
Take laboratory data set and consolidate Hain MTBDRplus results from Chechen data
}
\examples{
\dontrun{
mtbdrplus_fixer(p, set = "chechnya")
}
}
\seealso{
\code{\link{tbcleanr}}
}
\author{
Jay Achar \email{jay.achar@doctors.org.uk}
}
