% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nse_renamer.R
\name{nse_renamer}
\alias{nse_renamer}
\title{Non-Standard Evaluation variable renamer}
\usage{
nse_renamer(x, software = c("excel", "koch_6", "epiinfo"), project = c("kk",
  "chechnya"), file = c("adm", "lab", "clinical_lab"), fun = c("baseline",
  "converter", "drug_duration", "dst_baseliner", "ecg_collectr"), ...)
}
\arguments{
\item{x}{data frame}

\item{software}{define software used for data collection.
Values can be "excel", "koch_6", "epiinfo"}

\item{project}{define project location to apply.
Values can be "kk", "chechnya".}

\item{file}{define database file argument to apply.
Values can be "adm", "lab", "clinical_lab",}

\item{fun}{string to define which function requires NSE renaming}

\item{...}{further arguments passed to or from other methods}
}
\description{
Rename variables to standard sets to allow use of NSE packages
}
\examples{
\dontrun{
nse_renamer(p, software = "koch_6", project = "chechnya", file = "adm", fun = "drug_duration")
}
}
\seealso{
\code{\link{tbcleanr}}
}
\author{
Jay Achar \email{jay.achar@doctors.org.uk}
}
