% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oca_xpert_cleanr.R
\name{oca_xpert_cleanr}
\alias{oca_xpert_cleanr}
\title{Bentiu Xpert laboratory data cleaning}
\usage{
oca_xpert_cleanr(x)
}
\arguments{
\item{x}{data frome including Bentiu PoC lab Xpert data}
}
\description{
This function takes the Bentiu PoC lab dataset as a data.frame and
removes unnecessary variables, cleans results and generates new 
result variables.
}
\examples{
\dontrun{
oca_xpert_reporter(p)
}
}
\seealso{
\code{\link{tbcleanr}}
}
\author{
Jay Achar \email{jay.achar@doctors.org.uk}
}
