% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oca_xpert_date.R
\name{oca_xpert_date}
\alias{oca_xpert_date}
\title{Bentiu Xpert laboratory date adjustment}
\usage{
oca_xpert_date(x)
}
\arguments{
\item{x}{data frame including variable called "date"}
}
\description{
This function takes the Bentiu PoC lab dataset as a data.frame and cleans 
the date format and class.
}
\seealso{
\code{\link{tbcleanr}}
}
\author{
Jay Achar \email{jay.achar@doctors.org.uk}
}
\keyword{TB}
