% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outcome_fixer.default.R
\name{outcome_fixer.default}
\alias{outcome_fixer.default}
\title{Default method for outcome_fixer()}
\usage{
\method{outcome_fixer}{default}(x, ...)
}
\arguments{
\item{x}{data frame containing variables}

\item{...}{further arguments passed to or from other methods}
}
\description{
Allow data frames with unspecified object class to pass through
}
\seealso{
\code{\link{tbcleanr}}
}
\author{
Jay Achar
}
