% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outcome_fixer.koch6.R
\name{outcome_fixer.koch6}
\alias{outcome_fixer.koch6}
\title{Fix Koch 6 treatment outcomes}
\usage{
\method{outcome_fixer}{koch6}(x, who_defined = TRUE, rm_orig = TRUE,
  ...)
}
\arguments{
\item{x}{data frame containing outcome variables}

\item{who_defined}{create new factor variable with levels defined by WHO (2013)}

\item{rm_orig}{remove original variables - TRUE or FALSE}

\item{...}{further arguments passed to or from other methods}
}
\description{
Combine  treatment outcome variables to leave one 
factorised, labelled variable
}
\seealso{
\code{\link{tbcleanr}}
}
\author{
Jay Achar
}
