% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outcome_fixer.R
\name{outcome_fixer}
\alias{outcome_fixer}
\title{Fix Koch 6 treatment outcomes}
\usage{
outcome_fixer(x, software = c("excel", "koch_6", "epiinfo"), rm_orig = TRUE,
  ...)
}
\arguments{
\item{x}{data frame containing outcome variables}

\item{software}{define database being used - "koch_6", "epiinfo"}

\item{rm_orig}{remove original variables - TRUE or FALSE}

\item{...}{further arguments passed to or from other methods}
}
\description{
Combine  treatment outcome variables to leave one 
factorised, labelled variable
}
\examples{
\dontrun{
outcome_fixer(p, software = "epiinfo")
}
}
\seealso{
\code{\link{tbcleanr}}
}
\author{
Jay Achar \email{jay.achar@doctors.org.uk}
}
