% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recorded_dst.default.R
\name{recorded_dst.default}
\alias{recorded_dst.default}
\title{Default method for recorded_dst()}
\usage{
\method{recorded_dst}{default}(x)
}
\arguments{
\item{x}{data frame containing variables}
}
\description{
Allow data frames with unspecified object class to pass through
}
\seealso{
\code{\link{tbcleanr}}
}
\author{
Jay Achar
}
