% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recorded_dst.epiinfo.R
\name{recorded_dst.epiinfo}
\alias{recorded_dst.epiinfo}
\title{Clean recorded baseline DST variable}
\usage{
\method{recorded_dst}{epiinfo}(x)
}
\arguments{
\item{x}{data frame of TB admission data}
}
\value{
data frame with factorised baseline DST
}
\description{
Clean recorded baseline DST variable
}
