% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recorded_dst.koch6.R
\name{recorded_dst.koch6}
\alias{recorded_dst.koch6}
\title{Clean recorded baseline DST variable}
\usage{
\method{recorded_dst}{koch6}(x)
}
\arguments{
\item{x}{data frame of TB admission data}
}
\value{
data frame with factorised baseline DST
}
\description{
Clean recorded baseline DST variable
}
