% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recorded_dst.R
\name{recorded_dst}
\alias{recorded_dst}
\title{Clean recorded baseline DST variable}
\usage{
recorded_dst(x)
}
\arguments{
\item{x}{data frame of TB admission data}
}
\value{
data frame with factorised baseline DST
}
\description{
Clean recorded baseline DST variable
}
