% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/response_weight_cleanr.R
\name{response_weight_cleanr.default}
\alias{response_weight_cleanr.default}
\title{Treatment response weight cleanr}
\usage{
\method{response_weight_cleanr}{default}(x)
}
\arguments{
\item{x}{adherence data frame including monthly weight measurements}
}
\value{
data frame with treatment month and weight per patient. Duplicate
monthly measurements are removed with weights closest to the preceeding month
retained.
}
\description{
Treatment response weight cleanr
}
\seealso{
\code{\link{tbcleanr}}
}
\author{
Jay Achar
}
