% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/result_consolidator.R
\name{result_consolidator.default}
\alias{result_consolidator.default}
\title{Default method for result_consolidator()}
\usage{
\method{result_consolidator}{default}(x, rm_orig = TRUE, ...)
}
\arguments{
\item{x}{data frame containing variables}

\item{rm_orig}{remove original variables - TRUE or FALSE}

\item{...}{further arguments passed to or from other methods}
}
\description{
Allow data frames with unspecified object class to pass through
}
\seealso{
\code{\link{tbcleanr}}
}
\author{
Jay Achar
}
