% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/result_consolidator.epiinfo.R
\name{result_consolidator.epiinfo}
\alias{result_consolidator.epiinfo}
\title{Consolidate smear and culture results}
\usage{
\method{result_consolidator}{epiinfo}(x, rm_orig = TRUE, ...)
}
\arguments{
\item{x}{data frame containing variables}

\item{rm_orig}{remove original variables - TRUE or FALSE}

\item{...}{further arguments passed to or from other methods}
}
\description{
Take laboratory data set and consolidate repeated results to give summary variable
}
\seealso{
\code{\link{tbcleanr}}
}
\author{
Jay Achar \email{jay.achar@doctors.org.uk}
}
