% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/result_consolidator.R
\name{result_consolidator}
\alias{result_consolidator}
\title{Consolidate smear and culture results}
\usage{
result_consolidator(x, ...)
}
\arguments{
\item{x}{data frame containing variables}

\item{...}{further arguments passed to or from other methods}
}
\description{
Take laboratory data set and consolidate repeated results to give summary variable
}
\seealso{
\code{\link{tbcleanr}}
}
\author{
Jay Achar \email{jay.achar@doctors.org.uk}
}
