% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smear_recode.R
\name{smear_recode}
\alias{smear_recode}
\title{Recode smear results}
\usage{
smear_recode(sm)
}
\arguments{
\item{sm}{smear variable as defined in result_consolidator()}
}
\description{
Recode individual smear variables
}
\examples{
\dontrun{
smear_recode(sm)
}
}
\seealso{
\code{\link{TB.funs}}
}
\author{
Jay Achar \email{jay.achar@doctors.org.uk}
}
