% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subset_vars.R
\name{subset_vars}
\alias{subset_vars}
\title{Subset variables}
\usage{
subset_vars(x, set = "msc500")
}
\arguments{
\item{x}{data frame containing variables}

\item{set}{define variable set to apply. Values can be "msc500", "chechnya_myco_lab"}
}
\description{
Subset variables along pre-defined variable  
sets
}
\examples{
\dontrun{
subset_vars(p, set = "msc500")
}
}
\seealso{
\code{\link{TB.funs}}
}
\author{
Jay Achar \email{jay.achar@doctors.org.uk}
}
