% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TB.funs.R
\docType{package}
\name{TB.funs}
\alias{TB.funs}
\alias{TB.funs-package}
\title{TB.funs: A package for cleaning TB data}
\description{
The TB.funs package provides three important functions:
date.format, k6.clean.adm and tb.convert
}
\section{TB.funs functions}{

baseline
date.format
k6.clean.adm
tb.convert
}

