% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tb.convert.R
\name{tb_convert}
\alias{tb_convert}
\title{Calculate TB conversion date}
\usage{
tb_convert(x, id, date, start, result, days = 30)
}
\arguments{
\item{x}{data frame or data.table}

\item{id}{string representing patient unique ID number}

\item{date}{sample date (string)}

\item{start}{patient treatment start date (string)}

\item{result}{laboratory result variable (string)}

\item{days}{criteria for conversion}
}
\description{
This function takes long laboratory data and treatment start times to calculate culture conversion dates.
}
\keyword{TB}
