% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tbcleanr.R
\docType{package}
\name{tbcleanr}
\alias{tbcleanr}
\alias{tbcleanr-package}
\title{tbcleanr: A package for cleaning TB data.}
\description{
The tbcleanr package provides functions to clean and subset TB data from K6, and EpiInfo
}
\seealso{
\code{\link{adm_classr}}

\code{\link{adm_subset}}

\code{\link{adm_data_cleanr}}

\code{\link{binary_fixer}}

\code{\link{date_format}}

\code{\link{drug_fixer}}

\code{\link{gender_fixer}}

\code{\link{hiv_fixer}}

\code{\link{id_detangle}}

\code{\link{outcome_fixer}}

\code{\link{yn_binary_fixer}}
}
\author{
Jay Achar \email{jay.achar@doctors.org.uk}
}
