% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{transpose_tb_data}
\alias{transpose_tb_data}
\title{Clean and transpose Standard Indicators object}
\usage{
transpose_tb_data(df, year)
}
\arguments{
\item{df}{data frame of imported and extracted TB Standard Indicators}

\item{year}{character of length 1 representing the year of data collected}
}
\value{
data frame of tidy TB Standard Indicators data
}
\description{
Clean and transpose Standard Indicators object
}
