% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/txhistory.R
\name{txhistory.default}
\alias{txhistory.default}
\title{Default method for txhistory()}
\usage{
\method{txhistory}{default}(x)
}
\arguments{
\item{x}{data frame containing variables}
}
\description{
Allow data frames with unspecified object class to pass through
}
\seealso{
\code{\link{tbcleanr}}
}
\author{
Jay Achar \email{jay.achar@doctors.org.uk}
}
