% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/txhistory.epiinfo.R
\name{txhistory.epiinfo}
\alias{txhistory.epiinfo}
\title{Clean treatment history variable}
\usage{
\method{txhistory}{epiinfo}(x)
}
\arguments{
\item{x}{data frame of TB admission data}
}
\value{
data frame with factorised treatment history
variable
}
\description{
Clean treatment history variable
}
