% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/txhistory.koch6.R
\name{txhistory.koch6}
\alias{txhistory.koch6}
\title{Clean treatment history variable}
\usage{
\method{txhistory}{koch6}(x)
}
\arguments{
\item{x}{data frame of TB admission data}
}
\value{
data frame with factorised treatment history
variable
}
\description{
Clean treatment history variable
}
