% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/txhistory.R
\name{txhistory}
\alias{txhistory}
\title{Clean treatment history variable}
\usage{
txhistory(x)
}
\arguments{
\item{x}{data frame of TB admission data}
}
\value{
data frame with factorised treatment history
variable
}
\description{
Clean treatment history variable
}
