% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xpert_result_fixer.default.R
\name{xpert_result_fixer.default}
\alias{xpert_result_fixer.default}
\title{Default method for xpert_result_fixer()}
\usage{
\method{xpert_result_fixer}{default}(x, ...)
}
\arguments{
\item{x}{data frame containing admission ID variable}

\item{...}{further arguments passed to or from other methods}
}
\description{
Allow data frames with unspecified object class to pass through
}
\seealso{
\code{\link{tbcleanr}}
}
\author{
Jay Achar \email{jay.achar@doctors.org.uk}
}
