% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xpert_result_fixer.koch6.R
\name{xpert_result_fixer.koch6}
\alias{xpert_result_fixer.koch6}
\title{Consolidate Xpert results}
\usage{
\method{xpert_result_fixer}{koch6}(x, rm_orig = TRUE, ...)
}
\arguments{
\item{x}{data frame containing variables}

\item{rm_orig}{remove original variables - TRUE or FALSE}

\item{...}{further arguments passed to or from other methods}
}
\description{
Take laboratory data set and consolidate Xpert results
}
\seealso{
\code{\link{tbcleanr}}
}
\author{
Jay Achar
}
