% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xpert_result_fixer.R
\name{xpert_result_fixer}
\alias{xpert_result_fixer}
\title{Consolidate Xpert results}
\usage{
xpert_result_fixer(x, ...)
}
\arguments{
\item{x}{data frame containing variables}

\item{...}{further arguments passed to or from other methods}
}
\description{
Take laboratory data set and consolidate Xpert results
}
\seealso{
\code{\link{tbcleanr}}
}
\author{
Jay Achar \email{jay.achar@doctors.org.uk}
}
