% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{xpert_variable_detector}
\alias{xpert_variable_detector}
\title{Xpert variable detector}
\usage{
xpert_variable_detector(x)
}
\arguments{
\item{x}{data frame with xpert variables}
}
\value{
string of variable names matching regex for Xpert results
}
\description{
Find Xpert variables within data frame by using Regex. Adjust
when lab changes variable names
}
