% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/yn_binary_fixer.R
\name{yn_binary_fixer}
\alias{yn_binary_fixer}
\title{Factorise binary (Y/N) variables}
\usage{
yn_binary_fixer(x)
}
\arguments{
\item{x}{numerical variable with 2 levels - 0, 1}
}
\description{
Convert numerical variable into factor Y/N
}
\examples{
\dontrun{
yn_binary_fixer(p$renalfail)
}
}
\seealso{
\code{\link{TB.funs}}
}
\author{
Jay Achar \email{jay.achar@doctors.org.uk}
}
