% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zero_to_na.R
\name{zero_to_na}
\alias{zero_to_na}
\title{Convert zero's to NA in data frame}
\usage{
zero_to_na(x, software = c("excel", "koch_6", "epiinfo"),
  project = c("kk", "chechnya"), file = c("adm", "lab",
  "clinical_lab"), add = NULL, ...)
}
\arguments{
\item{x}{data frame containing variables}

\item{software}{define software used for data collection.
Values can be "excel", "koch_6", "epiinfo"}

\item{project}{define project location to apply.
Values can be "kk", "chechnya".}

\item{file}{define database file argument to apply.
Values can be "adm", "lab", "clinical_lab",}

\item{add}{string of any additional variables to include}

\item{...}{further arguments passed to or from other methods}
}
\description{
In specified data sets, convert all zero's to NAs in specific variables.
}
\examples{
\dontrun{
zero_to_na(z, set = "k6_clin_lab")
}
}
\seealso{
\code{\link{tbcleanr}}
}
\author{
Jay Achar \email{jay.achar@doctors.org.uk}
}
