% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adm_binary_fixer.R
\name{adm_binary_fixer}
\alias{adm_binary_fixer}
\title{Convert binary variables to factorised output}
\usage{
adm_binary_fixer(x, software = c("excel", "koch_6", "epiinfo"),
  project = c("kk", "chechnya"), file = c("adm", "lab", "clinical_lab"),
  ...)
}
\arguments{
\item{x}{data frame containing drug variables}

\item{software}{define software used for data collection.
Values can be "excel", "koch_6", "epiinfo"}

\item{project}{define project location to apply.
Values can be "kk", "chechnya".}

\item{file}{define database file argument to apply.
Values can be "adm", "lab", "clinical_lab",}

\item{...}{further arguments passed to or from other methods}
}
\description{
Take data frame with binary variables and output 
binary factorised variables
}
\examples{
\dontrun{
adm_binary_fixer(p, set = "msc500")
}
}
\seealso{
\code{\link{tbcleanr}}
}
\author{
Jay Achar \email{jay.achar@doctors.org.uk}
}
