% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adm_data_cleanr.R
\name{adm_data_cleanr}
\alias{adm_data_cleanr}
\title{Cleans TB admissions data sets}
\usage{
adm_data_cleanr(x, add = NULL, ...)
}
\arguments{
\item{x}{data frame containing variables}

\item{add}{string of additional variable names to retain in cleaned output data frame}

\item{...}{further arguments passed to or from other methods}
}
\description{
Take admission data set and perform multiple adjustments based on which
database set is being used
}
\seealso{
\code{\link{tbcleanr}}
}
\author{
Jay Achar \email{jay.achar@doctors.org.uk}
}
