% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adm_data_cleanr.R
\name{adm_data_cleanr}
\alias{adm_data_cleanr}
\title{Clean TB admission data sets}
\usage{
adm_data_cleanr(x, add = NULL, ...)
}
\arguments{
\item{x}{data frame containing variables}

\item{add}{string of additional variable names to retain in cleaned output data frame}

\item{...}{further arguments passed to or from other methods}
}
\value{
Data frame with an object attribute signifying the data collection software - 
"epiinfo" or "koch6"
}
\description{
Simplify workflow of cleaning TB admission data sets from EpiInfo or Koch 6
by adding an object class for identification, subsetting variables, 
extracting additional information from ID numbers in Epiinfo, formatting date, 
gender, HIV and treatment outcome variables. Also, converts drug dosing variables
to binary variables
}
\seealso{
\code{\link{tbcleanr}}
}
\author{
Jay Achar
}
