% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adm_data_cleanr.R
\name{adm_data_cleanr}
\alias{adm_data_cleanr}
\title{Cleans TB admissions data sets}
\usage{
adm_data_cleanr(x, set, ...)
}
\arguments{
\item{x}{data frame containing variables}

\item{set}{define admission data set being used. Values can be - "k6_adm_standard".}

\item{...}{further arguments passed to or from other methods}
}
\description{
Take admission data set and perform multiple adjustments based on which
database set is being used
}
\examples{
\dontrun{
adm_data_cleanr(p, lab = "chechnya")
}
}
\seealso{
\code{\link{tbcleanr}}
}
\author{
Jay Achar \email{jay.achar@doctors.org.uk}
}
