% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adm_subset.R
\name{adm_subset}
\alias{adm_subset}
\title{Subset adm variables}
\usage{
adm_subset(x, add = NULL, ...)
}
\arguments{
\item{x}{data frame containing variables}

\item{add}{string of any additional variables to keep}

\item{...}{further arguments passed to or from other methods}
}
\description{
Subset pre-specified TB admission variables
}
\seealso{
\code{\link{tbcleanr}}
}
\author{
Jay Achar
}
