% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cavities_fixer.R
\name{cavities_fixer}
\alias{cavities_fixer}
\title{Fix and aggregate Koch 6 x-ray cavities variables}
\usage{
cavities_fixer(x, db = "k6", rm_orig = TRUE, ...)
}
\arguments{
\item{x}{data frame containing xray variable data}

\item{db}{define database being used - "k6", "epi_info"}

\item{rm_orig}{remove original variables - TRUE or FALSE}

\item{...}{further arguments passed to or from other methods}
}
\description{
Input data frame containing cavity variables, factorise 
and aggregate into new variable -cavity-
}
\examples{
\dontrun{
cavities_fixer(p, db = "epi_info", rm_orig = TRUE)
}
}
\seealso{
\code{\link{TB.funs}}
}
\author{
Jay Achar \email{jay.achar@doctors.org.uk}
}
