% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/change_cleanr.koch6.R
\name{change_cleanr.koch6}
\alias{change_cleanr.koch6}
\title{Clean drug change data}
\usage{
\method{change_cleanr}{koch6}(x)
}
\arguments{
\item{x}{data frame containing drug variables}
}
\description{
Take drug change data from routine TB programmes and clean for 
further analysis.
}
\seealso{
\code{\link{tbcleanr}}
}
\author{
Jay Achar \email{jay.achar@doctors.org.uk}
}
