% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/change_cleanr.R
\name{change_cleanr}
\alias{change_cleanr}
\title{Clean drug change data}
\usage{
change_cleanr(x)
}
\arguments{
\item{x}{data frame containing drug variables}
}
\description{
Take drug change data from routine TB programmes and clean for 
further analysis.
}
\seealso{
\code{\link{tbcleanr}}
}
\author{
Jay Achar
}
