% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/culture_recode.R
\name{culture_recode}
\alias{culture_recode}
\title{Recode culture results}
\usage{
culture_recode(cult, software = c("excel", "koch_6", "epiinfo"),
  project = c("kk", "chechnya"), file = c("adm", "lab", "clinical_lab"),
  ...)
}
\arguments{
\item{cult}{culture variable as defined in result_consolidator()}

\item{software}{define software used for data collection.
Values can be "excel", "koch_6", "epiinfo"}

\item{project}{define project location to apply.
Values can be "kk", "chechnya".}

\item{file}{define database file argument to apply.
Values can be "adm", "lab", "clinical_lab",}

\item{...}{further arguments passed to or from other methods}
}
\description{
Recode individual culture variables
}
\examples{
\dontrun{
culture_recode(cult)
}
}
\seealso{
\code{\link{tbcleanr}}
}
\author{
Jay Achar \email{jay.achar@doctors.org.uk}
}
