% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drug_fixer.R
\name{drug_fixer}
\alias{drug_fixer}
\title{Convert drug doses to binary output}
\usage{
drug_fixer(x, set, ...)
}
\arguments{
\item{x}{data frame containing drug variables}

\item{set}{define variable set to apply. Values can be "msc500",
"k6_adm_standard", "nukus_epi_info"}

\item{...}{further arguments passed to or from other methods.}
}
\description{
Take data frame with drug dosing variables and output 
binary factorised variables
}
\examples{
\dontrun{
drug_fixer(p, set = "msc500")
}
}
\seealso{
\code{\link{TB.funs}}
}
\author{
Jay Achar \email{jay.achar@doctors.org.uk}
}
