% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dst_consolidator.R
\name{dst_consolidator.default}
\alias{dst_consolidator.default}
\title{Default method for dst_consolidator()}
\usage{
\method{dst_consolidator}{default}(x, aggregate = FALSE, rm_orig = TRUE)
}
\arguments{
\item{x}{data frame containing variables}

\item{aggregate}{choose whether to aggregate to categories or retain all drug results}

\item{rm_orig}{remove original variables - TRUE or FALSE}
}
\description{
Allow data frames with unspecified object class to pass through
}
\seealso{
\code{\link{tbcleanr}}
}
\author{
Jay Achar
}
