% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dst_consolidator.R
\name{dst_consolidator}
\alias{dst_consolidator}
\title{Consolidate DST results}
\usage{
dst_consolidator(x, aggregate = FALSE, rm_orig = TRUE, ...)
}
\arguments{
\item{x}{data frame containing variables}

\item{aggregate}{choose whether to aggregate to categories or retain all drug results}

\item{rm_orig}{remove original variables - TRUE or FALSE}

\item{...}{further arguments passed to or from other methods}
}
\description{
Take laboratory data set and consolidate DST results - choose aggregate
for categorised results
}
\seealso{
\code{\link{tbcleanr}}
}
\author{
Jay Achar \email{jay.achar@doctors.org.uk}
}
