% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dstno_detangle.R
\name{dstno_detangle}
\alias{dstno_detangle}
\title{Detangle DST identification number}
\usage{
dstno_detangle(x, software = c("excel", "koch_6", "epiinfo"),
  project = c("kk", "chechnya"), file = c("adm", "lab", "clinical_lab"),
  ...)
}
\arguments{
\item{x}{data frame containing APID variable from KK programme}

\item{software}{define software used for data collection.
Values can be "excel", "koch_6", "epiinfo"}

\item{project}{define project location to apply.
Values can be "kk", "chechnya".}

\item{file}{define database file argument to apply.
Values can be "adm", "lab", "clinical_lab",}

\item{...}{further arguments passed to or from other methods}
}
\description{
Convert character DST number found in Chechen data to numerical
}
\examples{
\dontrun{
dstno_detangle(p)
}
}
\seealso{
\code{\link{tbcleanr}}
}
\author{
Jay Achar \email{jay.achar@doctors.org.uk}
}
