% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gender_fixer.R
\name{gender_fixer}
\alias{gender_fixer}
\title{Factorise gender variable}
\usage{
gender_fixer(x, software = c("excel", "koch_6", "epiinfo"), rm_orig = TRUE,
  ...)
}
\arguments{
\item{x}{data frame containing Koch 6 admission variables}

\item{software}{define database being used - "koch_6", "epiinfo"}

\item{rm_orig}{remove original variables - TRUE or FALSE}

\item{...}{further arguments passed to or from other methods}
}
\description{
Take gender variable from data frame and factorise. Use software
to define which data base is being used as the input
}
\examples{
\dontrun{
gender_fixer(p, software = "koch_6")
}
}
\seealso{
\code{\link{tbcleanr}}
}
\author{
Jay Achar \email{jay.achar@doctors.org.uk}
}
