% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gender_fixer.R
\name{gender_fixer}
\alias{gender_fixer}
\title{Factorise gender variable}
\usage{
gender_fixer(x, db = "k6", rm_orig = TRUE, ...)
}
\arguments{
\item{x}{data frame containing Koch 6 admission variables}

\item{db}{define database being used - "k6", "epi_info"}

\item{rm_orig}{remove original variables - TRUE or FALSE}

\item{...}{further arguments passed to or from other methods}
}
\description{
Take gender variable from data frame and factorise. Use db
to define which data base is being used as the input
}
\examples{
\dontrun{
gender_fixer(p, db = "k6")
}
}
\seealso{
\code{\link{tbcleanr}}
}
\author{
Jay Achar \email{jay.achar@doctors.org.uk}
}
