% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gender_fixer.R
\name{gender_fixer}
\alias{gender_fixer}
\title{Factorise gender variable}
\usage{
gender_fixer(x, ...)
}
\arguments{
\item{x}{data frame containing Koch 6 admission variables}

\item{...}{further arguments passed to or from other methods}
}
\description{
Take gender variable from data frame and factorise. Requires that variable names
have been pre-adjusted.
}
\examples{
# data from Koch 6 with variable names pre-cleaned

x <- tbcleanr:::gender_k6
gender_fixer(x)

}
\seealso{
\code{\link{tbcleanr}}
}
\author{
Jay Achar \email{jay.achar@doctors.org.uk}
}
