% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/id_detangle.R
\name{id_detangle}
\alias{id_detangle}
\title{Detangle ID number}
\usage{
id_detangle(x, software = c("excel", "koch_6", "epiinfo"), project = c("kk",
  "chechnya"), file = c("adm", "lab", "clinical_lab"), rm_orig = FALSE, ...)
}
\arguments{
\item{x}{data frame containing APID variable from KK programme}

\item{software}{define software being used}

\item{project}{define project location to apply.
Values can be "kk", "chechnya".}

\item{file}{define database file argument to apply.
Values can be "adm", "lab", "clinical_lab",}

\item{rm_orig}{remove original variables - TRUE or FALSE}

\item{...}{further arguments passed to or from other methods}
}
\description{
Handle TB data ID number based on database. If APID, split 
into 3 constituent parts to ascertain district, DS or DR TB 
categorisation and numerical ID number. If
}
\examples{
\dontrun{
apid_detangle(p)
}
}
\seealso{
\code{\link{tbcleanr}}
}
\author{
Jay Achar \email{jay.achar@doctors.org.uk}
}
