% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/id_detangle.R
\name{id_detangle}
\alias{id_detangle}
\title{Detangle ID number}
\usage{
id_detangle(x, db = "k6", rm_orig = TRUE, ...)
}
\arguments{
\item{x}{data frame containing APID variable from KK programme}

\item{db}{define database being used - "k6", "epi_info"}

\item{rm_orig}{remove original variables - TRUE or FALSE}

\item{...}{further arguments passed to or from other methods}
}
\description{
Handle TB data ID number based on database. If APID, split 
into 3 constituent parts to ascertain district, DS or DR TB 
categorisation and numerical ID number. If
}
\examples{
\dontrun{
apid_detangle(p)
}
}
\seealso{
\code{\link{TB.funs}}
}
\author{
Jay Achar \email{jay.achar@doctors.org.uk}
}
