% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lab_data_cleanr.R
\name{lab_data_cleanr}
\alias{lab_data_cleanr}
\title{Cleans TB laboratory data sets}
\usage{
lab_data_cleanr(x, lab = "chechnya", ...)
}
\arguments{
\item{x}{data frame containing variables}

\item{lab}{define laboratory dat set being used. Values can be - "chechnya".}

\item{...}{further arguments passed to or from other methods}
}
\description{
Take laboratory data set and perform multiple adjustments based on which
laboratory data set is being used
}
\examples{
\dontrun{
lab_data_cleanr(p, lab = "chechnya")
}
}
\seealso{
\code{\link{tbcleanr}}
}
\author{
Jay Achar \email{jay.achar@doctors.org.uk}
}
