% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lab_data_cleanr.R
\name{lab_data_cleanr}
\alias{lab_data_cleanr}
\title{Cleans laboratory data sets}
\usage{
lab_data_cleanr(x, software = c("excel", "koch_6", "epiinfo"),
  project = c("kk", "chechnya"), file = c("adm", "lab", "clinical_lab"),
  add = NULL, ...)
}
\arguments{
\item{x}{data frame containing variables}

\item{software}{define software used for data collection.
Values can be "excel", "koch_6", "epiinfo"}

\item{project}{define project location to apply.
Values can be "kk", "chechnya".}

\item{file}{define database file argument to apply.
Values can be "adm", "lab", "clinical_lab",}

\item{add}{string of any additional variables to keep}

\item{...}{further arguments passed to or from other methods}
}
\description{
Take laboratory data set and perform multiple adjustments based on which
laboratory data set is being used
}
\examples{
\dontrun{
lab_data_cleanr(p, lab = "chechnya")
}
}
\seealso{
\code{\link{tbcleanr}}
}
\author{
Jay Achar \email{jay.achar@doctors.org.uk}
}
