% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lab_date_consolidator.R
\name{lab_date_consolidator}
\alias{lab_date_consolidator}
\title{Aggregate dates in lab database}
\usage{
lab_date_consolidator(x, software = c("excel", "koch_6", "epiinfo"),
  project = c("kk", "chechnya"), file = c("adm", "lab", "clinical_lab"),
  ...)
}
\arguments{
\item{x}{data frame containing sample date variables}

\item{software}{define software used for data collection.
Values can be "excel", "koch_6", "epiinfo"}

\item{project}{define project location to apply.
Values can be "kk", "chechnya".}

\item{file}{define database file argument to apply.
Values can be "adm", "lab", "clinical_lab",}

\item{...}{further arguments passed to or from other methods}
}
\description{
Consolidate sample dates in mycobacterial lab database
}
\examples{
\dontrun{
lab_date_consolidator(p, set = "msc500")
}
}
\seealso{
\code{\link{tbcleanr}}
}
\author{
Jay Achar \email{jay.achar@doctors.org.uk}
}
