% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lab_longr.R
\name{lab_longr}
\alias{lab_longr}
\title{Convert lab data from wide to long}
\usage{
lab_longr(x, set, add = NULL, ...)
}
\arguments{
\item{x}{data frame containing variables}

\item{set}{define variable set to apply. Values can be "k6_clin_lab".}

\item{add}{string of any additional variables to include}

\item{...}{further arguments passed to or from other methods}
}
\description{
In specified data sets, convert laboratory results from wide to long format.
}
\examples{
\dontrun{
lab_longr(z, set = "k6_clin_lab")
}
}
\seealso{
\code{\link{tbcleanr}}
}
\author{
Jay Achar \email{jay.achar@doctors.org.uk}
}
