% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lab_sample_fixer.R
\name{lab_sample_fixer}
\alias{lab_sample_fixer}
\title{Lab sample fixer}
\usage{
lab_sample_fixer(x, software = c("excel", "koch_6", "epiinfo"),
  project = c("kk", "chechnya"), file = c("adm", "lab", "clinical_lab"),
  rm_orig = TRUE, ...)
}
\arguments{
\item{x}{data frame containing sample date variables}

\item{software}{define software used for data collection.
Values can be "excel", "koch_6", "epiinfo"}

\item{project}{define project location to apply.
Values can be "kk", "chechnya".}

\item{file}{define database file argument to apply.
Values can be "adm", "lab", "clinical_lab",}

\item{rm_orig}{remove original variables - TRUE or FALSE}

\item{...}{further arguments passed to or from other methods}
}
\description{
Consolidate mycobacterial sample types in Chechen lab data
}
\examples{
\dontrun{
lab_sample_fixer(p, rm_orig = TRUE)
}
}
\seealso{
\code{\link{tbcleanr}}
}
\author{
Jay Achar \email{jay.achar@doctors.org.uk}
}
