% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lab_sample_fixer.R
\name{lab_sample_fixer}
\alias{lab_sample_fixer}
\title{Lab sample fixer}
\usage{
lab_sample_fixer(x, set = "chechnya_myco_lab", rm_orig = TRUE, ...)
}
\arguments{
\item{x}{data frame containing sample date variables}

\item{set}{define variable set to apply. Values can be "chechnya_myco_lab".}

\item{rm_orig}{remove original variables - TRUE or FALSE}

\item{...}{further arguments passed to or from other methods}
}
\description{
Consolidate mycobacterial sample types in Chechen lab data
}
\examples{
\dontrun{
lab_sample_fixer(p, rm_orig = TRUE)
}
}
\seealso{
\code{\link{TB.funs}}
}
\author{
Jay Achar \email{jay.achar@doctors.org.uk}
}
