% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mtbdrplus_fixer.R
\name{mtbdrplus_fixer}
\alias{mtbdrplus_fixer}
\title{Consolidate Hain MTBDRplus results}
\usage{
mtbdrplus_fixer(x, set = "chechnya", ...)
}
\arguments{
\item{x}{data frame containing variables}

\item{set}{define variable set to apply. Values can be - "chechnya"}

\item{...}{further arguments passed to or from other methods}
}
\description{
Take laboratory data set and consolidate Hain MTBDRplus results from Chechen data
}
\examples{
\dontrun{
mtbdrplus_fixer(p, set = "chechnya")
}
}
\seealso{
\code{\link{TB.funs}}
}
\author{
Jay Achar \email{jay.achar@doctors.org.uk}
}
