% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outcome_fixer.R
\name{outcome_fixer}
\alias{outcome_fixer}
\title{Fix TB treatment outcomes}
\usage{
outcome_fixer(x, who_defined = TRUE, bin_outcome = TRUE,
  rm_orig = TRUE, ...)
}
\arguments{
\item{x}{data frame containing outcome variables}

\item{who_defined}{create new factor variable with levels defined by WHO (2013)}

\item{bin_outcome}{create new binary factor variable simplifying treatment outcome}

\item{rm_orig}{remove original variables - TRUE or FALSE}

\item{...}{further arguments passed to or from other methods}
}
\description{
Combine treatment outcome variables to leave one
factorised, labelled variable
}
\seealso{
\code{\link{tbcleanr}}
}
\author{
Jay Achar
}
