% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outcome_fixer.R
\name{outcome_fixer}
\alias{outcome_fixer}
\title{Fix Koch 6 treatment outcomes}
\usage{
outcome_fixer(x, db = "k6", rm_orig = TRUE, ...)
}
\arguments{
\item{x}{data frame containing outcome variables}

\item{db}{define database being used - "k6", "epi_info"}

\item{rm_orig}{remove original variables - TRUE or FALSE}

\item{...}{further arguments passed to or from other methods}
}
\description{
Combine  treatment outcome variables to leave one 
factorised, labelled variable
}
\examples{
\dontrun{
outcome_fixer(p, db = "epi_info")
}
}
\seealso{
\code{\link{TB.funs}}
}
\author{
Jay Achar \email{jay.achar@doctors.org.uk}
}
