% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/result_consolidator.R
\name{result_consolidator}
\alias{result_consolidator}
\title{Consolidate smear results}
\usage{
result_consolidator(x, set = "chechnya_myco_lab", rm_orig = TRUE)
}
\arguments{
\item{x}{data frame containing variables}

\item{set}{define variable set to apply. Values can be - "chechnya_myco_lab"}

\item{rm_orig}{remove original variables - TRUE or FALSE}
}
\description{
Take laboratory data set and consolidate repeated results to give summary variable
}
\examples{
\dontrun{
result_consolidator(p, set = "chechnya_myco_lab", rm_orig = TRUE)
}
}
\seealso{
\code{\link{TB.funs}}
}
\author{
Jay Achar \email{jay.achar@doctors.org.uk}
}
