% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subset_vars.R
\name{subset_vars}
\alias{subset_vars}
\title{Subset variables}
\usage{
subset_vars(x, software = c("excel", "koch_6", "epiinfo"),
  project = c("kk", "chechnya"), file = c("adm", "lab",
  "clinical_lab"), add = NULL, ...)
}
\arguments{
\item{x}{data frame containing variables}

\item{software}{define software used for data collection.
Values can be "excel", "koch_6", "epiinfo"}

\item{project}{define project location to apply.
Values can be "kk", "chechnya".}

\item{file}{define database file argument to apply.
Values can be "adm", "lab", "clinical_lab",}

\item{add}{string of any additional variables to keep}

\item{...}{further arguments passed to or from other methods}
}
\description{
Subset variables along pre-defined variable  
sets
}
\examples{
\dontrun{
subset_vars(p, set = "msc500")
}
}
\seealso{
\code{\link{tbcleanr}}
}
\author{
Jay Achar \email{jay.achar@doctors.org.uk}
}
