% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TB.funs.R
\docType{package}
\name{TB.funs}
\alias{TB.funs}
\alias{TB.funs-package}
\title{TB.funs: A package for cleaning TB data.}
\description{
The TB.funs package provides functions to clean and subset TB data from K6, and EpiInfo
}
\seealso{
\code{\link{baseline}}

\code{\link{date.format}}

\code{\link{k6.clean.adm}}

\code{\link{tb.convert}}

\code{\link{bentiu_xpert_date}}
}
\author{
Jay Achar \email{jay.achar@doctors.org.uk}
}
