% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TB.funs.R
\docType{package}
\name{TB.funs}
\alias{TB.funs}
\alias{TB.funs-package}
\title{TB.funs: A package for cleaning TB data.}
\description{
The TB.funs package provides functions to clean and subset TB data from K6, and EpiInfo
}
\seealso{
\code{\link{adm_binary_fixer}}

\code{\link{adm_data_cleanr}}

\code{\link{age_generator}}

\code{\link{apid_detangle}}

\code{\link{baseline}}

\code{\link{bentiu_xpert_date}}

\code{\link{bmi_generator}}

\code{\link{chechnya_lab_date_consolidator}}

\code{\link{culture_recode}}

\code{\link{date_format}}

\code{\link{drug_fixer}}

\code{\link{dst_consolidator}}

\code{\link{dst_recoder}}

\code{\link{dstno_detangle}}

\code{\link{gender_fixer}}

\code{\link{hiv_fixer}}

\code{\link{k6.clean.adm}}

\code{\link{k6_cavities_fixer}}

\code{\link{k6_hep_fixer}}

\code{\link{k6_idno_fixer}}

\code{\link{k6_outcome_fixer}}

\code{\link{lab_data_cleanr}}

\code{\link{lab_sample_fixer}}

\code{\link{mtbdrplus_fixer}}

\code{\link{result_consolidator}}

\code{\link{smear_recode}}

\code{\link{subset_vars}}

\code{\link{tb.convert}}

\code{\link{xpert_result_fixer}}

\code{\link{yn_binary_fixer}}
}
\author{
Jay Achar \email{jay.achar@doctors.org.uk}
}
