% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tbcleanr.R
\docType{package}
\name{tbcleanr}
\alias{tbcleanr}
\alias{tbcleanr-package}
\title{tbcleanr: tools for cleaning TB data}
\description{
The tbcleanr package contains functions to clean and subset TB data from 
Koch 6 and EpiInfo used by MSF OCA. 

When published, it will integrate with tbgeneratr, tbcheckr, tbreportr
}
\seealso{
\code{\link{adhere_classr}}

\code{\link{adm_classr}}

\code{\link{adm_subset}}

\code{\link{adm_data_cleanr}}

\code{\link{binary_fixer}}

\code{\link{cavities_fixer}}

\code{\link{change_cleanr}}

\code{\link{clean_standard_indicators}}

\code{\link{date_format}}

\code{\link{dst_consolidator}}

\code{\link{drug_fixer}}

\code{\link{gender_fixer}}

\code{\link{hiv_fixer}}

\code{\link{id_detangle}}

\code{\link{lab_classr}}

\code{\link{lab_data_cleanr}}

\code{\link{lab_import_checkr}}

\code{\link{lab_sample_fixer}}

\code{\link{lab_subset}}

\code{\link{mtbdrplus_fixer}}

\code{\link{outcome_fixer}}

\code{\link{recorded_dst}}

\code{\link{response_weight_cleanr}}

\code{\link{result_consolidator}}

\code{\link{txhistory}}

\code{\link{xpert_result_fixer}}

\code{\link{yn_binary_fixer}}
}
\author{
Jay Achar \email{jay.achar@doctors.org.uk}
}
