% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xpert_result_fixer.R
\name{xpert_result_fixer}
\alias{xpert_result_fixer}
\title{Consolidate Xpert results}
\usage{
xpert_result_fixer(x, software = c("excel", "koch_6", "epiinfo"),
  project = c("kk", "chechnya"), file = c("adm", "lab",
  "clinical_lab"), rm_orig = TRUE, ...)
}
\arguments{
\item{x}{data frame containing variables}

\item{software}{define software used for data collection.
Values can be "excel", "koch_6", "epiinfo"}

\item{project}{define project location to apply.
Values can be "kk", "chechnya".}

\item{file}{define database file argument to apply.
Values can be "adm", "lab", "clinical_lab",}

\item{rm_orig}{remove original variables - TRUE or FALSE}

\item{...}{further arguments passed to or from other methods}
}
\description{
Take laboratory data set and consolidate Xpert results from Chechen data
}
\examples{
\dontrun{
xpert_result_fixer(p, set = "chechnya_myco_lab", rm_orig = TRUE)
}
}
\seealso{
\code{\link{tbcleanr}}
}
\author{
Jay Achar \email{jay.achar@doctors.org.uk}
}
