% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xpert_result_fixer.R
\name{xpert_result_fixer}
\alias{xpert_result_fixer}
\title{Consolidate Xpert results}
\usage{
xpert_result_fixer(x, set = "chechnya_myco_lab", rm_orig = TRUE)
}
\arguments{
\item{x}{data frame containing variables}

\item{set}{define variable set to apply. Values can be - "chechnya_myco_lab"}

\item{rm_orig}{remove original variables - TRUE or FALSE}
}
\description{
Take laboratory data set and consolidate Xpert results from Chechen data
}
\examples{
\dontrun{
xpert_result_fixer(p, set = "chechnya_myco_lab", rm_orig = TRUE)
}
}
\seealso{
\code{\link{TB.funs}}
}
\author{
Jay Achar \email{jay.achar@doctors.org.uk}
}
